core: extract shared summary/protocol into oura-summary + oura-core/oura-ffi
Move build_summary and its helpers out of the CLI into a reusable `oura-summary`
crate behind a `ModelRunner` trait (the dashboard supplies a Python runner; a native
client can supply an on-device one). Add `oura-core` (UniFFI-exported protocol/summary
surface) and `oura-ffi` (C-ABI) so the same tested Rust logic backs the native app.
dashboard.rs now just wires the Python runner + HTTP server; no behavior change.
