+++ crates/oura-cli/Cargo.toml
@@ -14,6 +14,7 @@ path = "src/main.rs"
 oura-protocol = { path = "../oura-protocol" }
 oura-link = { path = "../oura-link" }
 oura-store = { path = "../oura-store" }
+oura-summary = { path = "../oura-summary" }
 clap = { version = "4", features = ["derive"] }
 tokio = { version = "1", features = ["full"] }
 anyhow = "1"
