+++ crates/oura-core/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "oura-core"
+description = "UniFFI surface over the shared Oura core — consumed by the iOS/native clients"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[lib]
+name = "oura_core"
+# lib for the host bins; staticlib for iOS linking; cdylib for library-mode bindgen.
+crate-type = ["lib", "staticlib", "cdylib"]
+
+[dependencies]
+oura-analysis = { path = "../oura-analysis" }
+oura-store = { path = "../oura-store" }
+oura-summary = { path = "../oura-summary" }
+uniffi = { version = "0.28", features = ["cli"] }
+serde_json = "1"
+
+[[bin]]
+# `cargo run --bin uniffi-bindgen -- generate --library <dylib> --language swift ...`
+name = "uniffi-bindgen"
+path = "uniffi-bindgen.rs"
