+++ crates/oura-link/Cargo.toml
@@ -14,4 +14,10 @@ uuid = "1"
 tokio = { version = "1", features = ["sync", "time", "rt", "macros"] }
 async-trait = "0.1"
 futures = "0.3"
-btleplug = "0.11"
+# BLE backend. Optional so the auth + sync logic (OuraClient, transport-generic)
+# can build for iOS, where CoreBluetooth implements the `Transport` trait instead.
+btleplug = { version = "0.11", optional = true }
+
+[features]
+default = ["ble"]
+ble = ["dep:btleplug"]
